# [timeseries] Speed up featurization for AutoGluonTabularModel