--- a/timeseries/tests/unittests/models/test_autogluon_tabular.py
+++ b/timeseries/tests/unittests/models/test_autogluon_tabular.py
@@ -15,18 +15,20 @@ TESTABLE_MODELS = [
 
 
 @pytest.mark.parametrize(
-    "data, last_k_values, expected_length",
+    "data, max_rows_per_item, expected_length",
     [
         (DUMMY_VARIABLE_LENGTH_TS_DATAFRAME, None, DUMMY_VARIABLE_LENGTH_TS_DATAFRAME.shape[0]),
         (DUMMY_VARIABLE_LENGTH_TS_DATAFRAME, 1, DUMMY_VARIABLE_LENGTH_TS_DATAFRAME.num_items),
         (DUMMY_VARIABLE_LENGTH_TS_DATAFRAME, 3, 3 * DUMMY_VARIABLE_LENGTH_TS_DATAFRAME.num_items),
     ],
 )
-def test_when_feature_df_is_constructed_then_shape_is_correct(data, last_k_values, expected_length, temp_model_path):
+def test_when_feature_df_is_constructed_then_shape_is_correct(
+    data, max_rows_per_item, expected_length, temp_model_path
+):
     model = AutoGluonTabularModel(path=temp_model_path)
     # Initialize model._lag_indices and model._time_features from freq
     model.fit(train_data=data, time_limit=2)
-    df = model._get_features_dataframe(data, last_k_values=last_k_values)
+    df = model._get_features_dataframe(data, max_rows_per_item=max_rows_per_item)
     expected_num_features = len(model._target_lag_indices) + len(model._time_features) + 1
     assert df.shape == (expected_length, expected_num_features)