*Description of changes:*
- Speed up the `AutoGluonTabularModel._get_features_dataframe` method using the following 3 improvements:
  1. Parallelize the construction of the feature matrix across individual time series with `joblib`.
  2. Initialize the full matrix as a numpy array and fill its values (instead of concatenating pd.DataFrames)
  3. During prediction, the old code used to construct the full matrix and then select the last `prediction_length` rows; now we only construct the last `prediction_length` rows.

**Checking correctness:**
- I checked that both old & new `_get_features_dataframe` methods produce identical results on synthetic datasets generated by taking the cartesian product of these configs:
```python
"freq", ["30S", "T", "A", "Y", "M", "H"]
"use_past_covariates", [True, False]
"use_known_covariates", [True, False]
"use_static_features_continuous", [True, False]
"use_static_features_categorical", [True, False]
"prediction_length", [1, 5]
"max_len", [13, 100_000]
```
**Performance comparison**
Comparing the runtime for the `AutoGluonTabular` model (using only the `LightGBM` model under the hood).
| Dataset     | Fit time old | Fit time new | Pred time old (s) | Pred time new (s)| Number of rows | Number of items |
|-------------|--------------|--------------|---------------|---------------|----------------|-----------------|
| M3 other    | 16.3         | 15.7         | 1.7           | 0.8           | 12K            | 174             |
| M4 hourly   | 49.3         | 34.5         | 10.8          | 2.2           | 353K           | 414             |
| Electricity | 261.9        | 242.7        | 38.2          | 12.7          | 6755K          | 321             |
| M4 daily    | 246.4        | 96.6         | 116.4         | 33.9          | 9964K          | 4227            |
|M5 (with `sell_price` past covariate & 5 static features)              | 2524.5      | 433.9      | 1829.2       | 673.7        | 46796K        | 30490         |

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.